fix: DELETE error with ObjectId

+++ route/upload/controller.js
@@ -1,4 +1,5 @@
 const CreateError = require('http-errors')
+const mongoose = require('mongoose')
 const signale = require('signale')
 const Qiniu = require('../../util/qiniu')
 
@@ -81,16 +82,20 @@ const getFile = db => async (req, reply) => {
 // 从mongodb删除指定文件
 const deleteFile = db => (req, reply) => {
   const { id } = req.params
-  db.findOneAndDelete({ _id: id })
-    .then(async (data) => {
-      try {
-        await qiniu.deleteFile(data.name)
-        reply.code(204).send()
-      } catch (err) {
-        throw err
-      }
-    })
-    .catch(err => reply.send(err))
+  /* eslint-disable */
+  if (!!mongoose.Types.ObjectId.isValid(id)) { 
+    return db.findOneAndDelete({ _id: id })
+      .then(async (data) => {
+        try {
+          await qiniu.deleteFile(data.name)
+          reply.code(204).send()
+        } catch (err) {
+          throw err
+        }
+      })
+      .catch(err => reply.send(err))
+  }
+  return reply.send(new CreateError.BadRequest('无效的id'))
 }
 
 module.exports = {
